1.消除 warning 2.优化聊天的图片浏览器

+++ PeachTravel/MJPhotoBrowser/MJPhotoToolbar.m
@@ -65,7 +65,6 @@
 - (void)image:(UIImage *)image didFinishSavingWithError:(NSError *)error contextInfo:(void *)contextInfo
 {
     if (error) {
-//        [SVProgressHUD showSuccess:@"保存失败" toView:nil];
         [SVProgressHUD showHint:@"保存失败"];
     } else {
         MJPhoto *photo = _photos[_currentPhotoIndex];
@@ -80,7 +79,7 @@
     _currentPhotoIndex = currentPhotoIndex;
     
     // 更新页码
-    _indexLabel.text = [NSString stringWithFormat:@"%d / %d", _currentPhotoIndex + 1, _photos.count];
+    _indexLabel.text = [NSString stringWithFormat:@"%lu / %lu", (long)_currentPhotoIndex + 1, (unsigned long)_photos.count];
     
     MJPhoto *photo = _photos[_currentPhotoIndex];
     // 按钮
